+++ README.md
@@ -21,3 +21,5 @@
 9. ESLint: We want our code to look nice and consistent. Let's add some linting rules!
  
 10. Some refactoring: fetchBalances and helper functions should go to a separate file. That would also simplify testing it.
+
+11. Add useFetchBalances hook, that will wrap fetchBalances function and return loading, error and balances. That will simplify usage of fetchBalances function and App would display proper states
